moved Criterion into the query package
